Extract shared CTA banner into SharedPageCtaBanner component
Both shared public pages (review and watchlist) duplicated the same
fixed-bottom banner and scroll-tracking logic. Extracts into a single
self-contained component.
