+++ src/features/reviews/views/SharedReviewView.vue
@@ -124,78 +124,28 @@
       />
     </div>
 
-    <!-- Fixed Call-to-Action Banner -->
-    <div v-if="!isLoggedIn">
-      <div class="h-20" />
-      <div
-        class="fixed inset-x-0 bottom-0 bg-secondary px-4 py-3 shadow-lg transition-transform duration-300"
-        :class="{ 'translate-y-full': isScrollingDown }"
-      >
-        <div class="mx-auto flex max-w-6xl items-center justify-between gap-4">
-          <div class="text-left">
-            <h2 class="text-lg font-bold text-white">Join the Club!</h2>
-            <p class="text-xs text-gray-200">
-              Join clubs and review movies with your friends.
-            </p>
-          </div>
-          <a
-            href="/"
-            class="whitespace-nowrap rounded-md bg-primary px-4 py-2 text-sm font-semibold text-white transition hover:bg-primary/80"
-          >
-            Sign Up
-          </a>
-        </div>
-      </div>
-    </div>
+    <SharedPageCtaBanner />
   </div>
 </template>
 
 <script setup lang="ts">
 import { DateTime } from "luxon";
-import { computed, ref, onMounted, onUnmounted } from "vue";
+import { computed } from "vue";
 import { useRoute } from "vue-router";
 
 import { hasElements } from "../../../../lib/checks/checks.js";
 
 import LoadingSpinner from "@/common/components/LoadingSpinner.vue";
+import SharedPageCtaBanner from "@/common/components/SharedPageCtaBanner.vue";
 import VAvatar from "@/common/components/VAvatar.vue";
 import SharedReviewComments from "@/features/reviews/components/SharedReviewComments.vue";
 import { useSharedReview } from "@/service/useList";
-import { useAuthStore } from "@/stores/auth";
 
 const route = useRoute();
 const clubSlug = route.params.clubSlug as string;
 const workId = route.params.workId as string;
 
-const authStore = useAuthStore();
-const isLoggedIn = computed(() => authStore.isLoggedIn);
-
 const { data, isLoading, error } = useSharedReview(clubSlug, workId);
-
-// Scroll behavior for floating bar
-const isScrollingDown = ref(false);
-let lastScrollY = 0;
-
-const handleScroll = () => {
-  const currentScrollY = window.scrollY;
-
-  // Only hide if scrolling down and scrolled at least 50px
-  if (currentScrollY > lastScrollY && currentScrollY > 50) {
-    isScrollingDown.value = true;
-  } else if (currentScrollY < lastScrollY) {
-    isScrollingDown.value = false;
-  }
-
-  lastScrollY = currentScrollY;
-};
-
-onMounted(() => {
-  window.addEventListener("scroll", handleScroll, { passive: true });
-});
-
-onUnmounted(() => {
-  window.removeEventListener("scroll", handleScroll);
-});
 
 const averageScore = computed(() => {
   if (!data.value) return 0;
