+++ src/features/watch-list/views/SharedWatchlistView.vue
@@ -44,34 +44,12 @@
       </div>
     </div>
 
-    <!-- Fixed Call-to-Action Banner -->
-    <div v-if="!isLoggedIn">
-      <div class="h-20" />
-      <div
-        class="fixed inset-x-0 bottom-0 bg-secondary px-4 py-3 shadow-lg transition-transform duration-300"
-        :class="{ 'translate-y-full': isScrollingDown }"
-      >
-        <div class="mx-auto flex max-w-6xl items-center justify-between gap-4">
-          <div class="text-left">
-            <h2 class="text-lg font-bold text-white">Join the Club!</h2>
-            <p class="text-xs text-gray-200">
-              Join clubs and review movies with your friends.
-            </p>
-          </div>
-          <a
-            href="/"
-            class="whitespace-nowrap rounded-md bg-primary px-4 py-2 text-sm font-semibold text-white transition hover:bg-primary/80"
-          >
-            Sign Up
-          </a>
-        </div>
-      </div>
-    </div>
+    <SharedPageCtaBanner />
   </div>
 </template>
 
 <script setup lang="ts">
-import { computed, onMounted, onUnmounted, ref } from "vue";
+import { computed } from "vue";
 import { useRoute } from "vue-router";
 
 import { hasElements } from "../../../../lib/checks/checks.js";
@@ -79,15 +57,12 @@ import { WorkListType } from "../../../../lib/types/generated/db.js";
 
 import EmptyState from "@/common/components/EmptyState.vue";
 import MoviePosterCard from "@/common/components/MoviePosterCard.vue";
+import SharedPageCtaBanner from "@/common/components/SharedPageCtaBanner.vue";
 import { useClub } from "@/service/useClub";
 import { useList, useNextWork } from "@/service/useList";
-import { useAuthStore } from "@/stores/auth";
 
 const route = useRoute();
 const clubSlug = route.params.clubSlug as string;
-
-const authStore = useAuthStore();
-const isLoggedIn = computed(() => authStore.isLoggedIn);
 
 const { data: club } = useClub(clubSlug);
 const {
@@ -104,27 +79,5 @@ const sortedWatchList = computed(() => {
     return [nextItem, ...list.filter((item) => item.id !== nextItem.id)];
   }
   return list;
-});
-
-// Scroll behavior for floating CTA bar
-const isScrollingDown = ref(false);
-let lastScrollY = 0;
-
-const handleScroll = () => {
-  const currentScrollY = window.scrollY;
-  if (currentScrollY > lastScrollY && currentScrollY > 50) {
-    isScrollingDown.value = true;
-  } else if (currentScrollY < lastScrollY) {
-    isScrollingDown.value = false;
-  }
-  lastScrollY = currentScrollY;
-};
-
-onMounted(() => {
-  window.addEventListener("scroll", handleScroll, { passive: true });
-});
-
-onUnmounted(() => {
-  window.removeEventListener("scroll", handleScroll);
 });
 </script>
